refactor: Remove B2BUnitNode model from core lib
