+++ feature-libs/my-account/organization/components/permission/form/permission-form.component.ts
@@ -1,7 +1,8 @@
 import { ChangeDetectionStrategy, Component, OnInit } from '@angular/core';
 import { FormGroup } from '@angular/forms';
-import { B2BUnitNode, Currency, CurrencyService } from '@spartacus/core';
+import { Currency, CurrencyService } from '@spartacus/core';
 import {
+  B2BUnitNode,
   OrderApprovalPermissionType,
   OrgUnitService,
   Period,
