+++ feature-libs/my-account/organization/components/unit/services/unit-list.service.ts
@@ -1,6 +1,9 @@
 import { Injectable } from '@angular/core';
-import { B2BUnit, B2BUnitNode, EntitiesModel } from '@spartacus/core';
-import { OrgUnitService } from '@spartacus/my-account/organization/core';
+import { B2BUnit, EntitiesModel } from '@spartacus/core';
+import {
+  B2BUnitNode,
+  OrgUnitService,
+} from '@spartacus/my-account/organization/core';
 import { TableService } from '@spartacus/storefront';
 import { Observable } from 'rxjs';
 import { map } from 'rxjs/operators';
