+++ feature-libs/my-account/organization/core/model/index.ts
@@ -1,5 +1,6 @@
 export * from './budget.model';
 export * from './order-approval.model';
 export * from './permission.model';
+export * from './unit-node.model';
 export * from './user-group.model';
 export * from './user.model';
