+++ feature-libs/my-account/organization/core/model/unit-node.model.ts
@@ -1,0 +1,7 @@
+export interface B2BUnitNode {
+  active?: boolean;
+  children?: B2BUnitNode[];
+  id?: string;
+  name?: string;
+  parent?: string;
+}
