+++ feature-libs/my-account/organization/core/store/organization-state.ts
@@ -2,7 +2,6 @@ import {
   B2BAddress,
   B2BApprovalProcess,
   B2BUnit,
-  B2BUnitNode,
   B2BUser,
   CostCenter,
   ListModel,
@@ -14,6 +13,7 @@ import {
   OrderApprovalPermissionType,
   Permission,
 } from '../model/permission.model';
+import { B2BUnitNode } from '../model/unit-node.model';
 import { UserGroup } from '../model/user-group.model';
 
 export const ORGANIZATION_FEATURE = 'organization';
